M.I.D.A.S/Backend/requirements.txt
asgiref==3.7.2
Django==5.0.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
python-dotenv==0.19.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
